refactor: rename back to FeatureCards
