+++ src/components/FeatureCards/index.ts
@@ -1,4 +1,4 @@
 export * from '../FeatureCards/FeatureCard';
 export * from '../FeatureCards/FeatureCard.style';
-export * from './FeatureCardsWrapper';
-export * from './FeatureCardsWrapper.style';
+export * from './FeatureCards';
+export * from './FeatureCards.style';
